python3/studys/230602/requirements.txt
yaml
xlrd == 1.2.0
jpype1
pycryptodome
